+++ frontend/src/components/@common/progressBar/linear/LinearProgressBar.styles.ts
