+++ frontend/src/components/@common/progressBar/step/StepProgressBar.styles.ts
