+++ frontend/src/components/@common/progressBar/step/StepProgressBar.tsx
