+++ frontend/src/components/specific/comingSoonBox/ComingSoonBox.styles.ts
