+++ frontend/src/components/specific/comingSoonBox/ComingSoonBox.tsx
