+++ frontend/src/components/specific/dashboardBox/DashboardBox.styles.ts
